Revisions to the shoot axis internode morphology trait (TO:0000755)
Was: shoot internode morphology trait (TO:0000755): A cardinal organ part anatomy and morphology trait (TO:0000754) which is associated with a shoot internode (PO_0005005).

Revised: shoot axis internode morphology trait (TO:0000755):
A cardinal organ part morphology trait (TO:0000754) which is a quality of a shoot axis internode (PO:0005005).

TO_GIT:64
_____________________________________________

stem internode morphology trait (TO:0000756): A shoot internode morphology trait (TO:0000755) which is associated with a stem internode (PO:0020142).

Revised:  A shoot axis internode morphology trait (TO:0000755) which is quality of a stem internode (PO:0020142).

#64