Revised ear infructescence height terms: Was: ear infructescence height (TO:0000683): A height related trait (TO:0000171) which is the height of an ear infructescence (PO:0025597) measured from the ground to the attachment point.
Revised name and definition:
ear infructescence position (TO:0000683): An ear infructescence morphology trait (TO:0000976) which is the position of an ear infructescence (PO:0025597).
Comment: Refers to: position (PATO_0000140): A spatial quality inhering in a bearer by virtue of the bearer's spatial location relative to other objects in the vicinity.  In maize this is usually measured from the ground to the attachment point of the ear.
Revised name and definition:
ear position (TO:0000966): An ear infructescence morphology trait (TO:0000976) which is the ratio of ear height to plant height.

ear infructescence position ratio (TO:0000966): An ear infructescence morphology trait (TO:0000976) which is the ratio of ear infructescence position (TO:0000683) to plant height (TO:0000207).

#132