Children of whole plant morphology trait (TO:0000398):
Moved plant height (TO:0000207) to be a child of whole plant morphology trait (TO:0000398).

Was:
plant height (TO:0000207): A height trait (TO:0000171) which is the height of a whole plant (PO:0000003).
Comment: Refers to: height (PATO_0000119): A 1-D extent quality inhering in a bearer by virtue of the bearer's vertical dimension of extension.

Revised:
plant height (TO:0000207): A whole plant morphology trait (TO:0000398) which is the height of a whole plant (PO:0000003).
Comment: Refers to: height (PATO_0000119): A 1-D extent quality inhering in a bearer by virtue of the bearer's vertical dimension of extension.

#84; #367