Moved plant color (TO:0000708) to be a child of whole plant morphology trait (TO:0000398).
Was: plant color (TO:0000708):  A stature and vigor trait (TO:0000133) which is the color of a whole plant (PO:0000003).

Revised: A whole plant morphology trait (TO:0000398) which is the color of a whole plant (PO:0000003).
Added comment: Refer to color (PATO:0000014): A composite chromatic quality composed of hue, saturation and intensity parts.

TO_GIT: 85; 367