Revised name and definiton from height to position * leaf height (TO:0000541): Was: leaf height (TO:0000541): A leaf morphology trait (TO:0000748) which is the height of the leaf (PO:0025034). Revised: leaf position (TO:0000541): A leaf morphology trait (TO:0000748) which is the position of the leaf (PO:0025034) on the plant. Comment: Refers to: position (PATO_0000140): A spatial quality inhering in a bearer by virtue of the bearer's spatial location relative to other objects in the vicinity. added related synonym: leaf height Revised equiv. axiom: Was: height and inheres_in some 'vascular leaf' Revised: position and (inheres_in some leaf)
Revised name and definition from height to position:
was:
* third leaf height  (TO:0000677): Height of the 3rd leaf measured as described in various studies.

third leaf position (TO:0000677): A leaf position (TO:0000541) which is the position of the third leaf (PO:0025034) on a plant.

Comment: Refers to: position (PATO_0000140): A spatial quality inhering in a bearer by virtue of the bearer's spatial location relative to other objects in the vicinity.
added related synonym: third leaf height