TO edits
* modified the equivalence axioms for the root term 'plant trait':
was: quality and ('occurs in' some 'whole plant')
Revised:
quality and ('occurs in' some 'plant anatomical entity' or 'plant structure development stage')

* revised term inserted by TermGenie:
stem internode bent (TO:1000000)
Revised equivalence axiom:
Was: PATO:0000617 and ('inheres in' some 'stem internode')

Changed to:  shape and ('inheres in' some 'stem internode')  Cannot use bent as it is not there.

* Fruit morphology trait:  Revised equivalence axiom:
was: morphology and ('inheres in' some fruit)
New: 'physical object quality' and ('inheres in' some fruit)

Revised name and definiton:
inflorescence height (TO:0000682): A height related trait (TO:0000171) which is associated with the height of an inflorescence (PO:0009049) measured from ground to the attachment point.
(TO_GIT: 366)

Revised name and definiton:
inflorescence position (TO:0000682): An inflorescence morphology trait (TO:0000373) which is the position of an inflorescence (PO:0009049) on a plant.

Comment: Refers to: position (PATO_0000140): A spatial quality inhering in a bearer by virtue of the bearer's spatial location relative to other objects in the vicinity. Usually measured from the ground.  This is often  measured from the ground to the point where the inflorescence is attached.
modified equivalence axiom:  position and (inheres_in some inflorescence)
#97 and  #226