height trait (TO:0000171): A stature and vigor trait (TO:0000133) which is the height of a whole plant (PO:0000003) or another plant structure (PO:0009011). Deprecating this grouping term as the child terms have been moved.
Revised name and def'n:
WAS:
- inflorescence height (TO:0000682):  An inflorescence morphology trait (TO:0000373) which is the height of an inflorescence (PO:0009049) measured from ground to the attachment point.
Revised:  inflorescence position (TO:0000682):  An inflorescence morphology trait (TO:0000373) which is the position of an inflorescence (PO:0009049) on the plant.

Comment: inflorescence position is usually measured from the ground to the attachment point.

Moved plant volume (TO:0000559), plant aspect (TO:0000737) and plant width (TO:0006062) to be children of whole plant morphology trait (TO:0000398).  Need to revise def'ns.

space per culm (TO:0000560) was a child of plant volume (TO:0000559)- moved to be is_a stem morphology trait (TO:0000361)
____________________________