Revised name and definiton: was "biological process related trait  (TO:0000283)"
New: biological process trait (TO:0000283):
A plant trait (TO:0000387) which is a process quality (PATO:0001236) of a biological process (GO:0008150) which occurs in a plant anatomical entity (PO:0025131), a cellular component (GO:0005575), or during a plant structure development stage (PO:0009012).

Added comment: Refers to process quality (PATO:0001236): A quality which inheres in an process.

Add EA: 'process quality' and 'inheres in' some biological_process