Revised def'ns and moved dbxref for CHEBI
amylose content  (TO:0000196):
Current def'n: Amount of amylose present in the plant or plant part.

Revised def'n: A polysaccharide content trait (TO:0006007) which is the amount of amylose (CHEBI:28102) present in a plant structure (PO:0009011)

Revised comment:
Was: It is a water-dispersible component of starch and is a polymer of alpha(1,4)-linked glucose units.

Revised: Amylose is a helical polymer made of α-D-glucose units, bound to each other through α(1→4) glycosidic bonds. This polysaccharide is one of the two components of starch, making up approximately 20-30% of the structure. The other component is amylopectin, which makes up 70–80% of the structure. (REF: Wikipedia)

is_a TO:0006007 polysaccharide content

Changed the subclass of relation to 'cooking or brewing quality' and 'carbohydrate composition related trait'  from is_a to part_of

 'part of' some 'cooking or brewing quality'
  'part of' some 'carbohydrate composition related trait'