Revised def'ns and comments,  and moved dbxref for CHEBI:
carotene content (TO:0000289):
Was:  The carotene content in a plant or plant part.

Revised def'n:
A carotenoid content trait (TO:0000496) which is the amount of carotene  (CHEBI:23042) present in a plant structure (PO:0009011).

Comment: 'These are any group of naturally occurring tetraterpenes (carotenoids) that are synthesized only in plants and differ from xanthophyll by not having oxygen in their molecule.'

Revised" Carotenes are a group of naturally occurring tetraterpenes (carotenoids) that are synthesized only in plants, and unlike the xanthophylls, do not contain any oxygen.

Moved CHEBI:23042 to dbxrefs from def'n dbxref
Added TO_GIT:24