Renamed lemma and palea morphology trait (TO:0000079) to be flower bract morphology trait (TO:0000079). to match the logical definition Revised def'n: Was: An inflorescence bract morphology trait (TO:0000816) which is a quality of a lemma (PO:0009037) and palea (PO:0009038) and their parts.
Revised: A bract morphology trait (TO:0000815)   which is a quality of a flower bract (PO:0009034).

Moved apiculus color (TO:0000140) and apiculus hair length (TO:0000532)- to be is_a children of cardinal organ part morphology trait (TO:0000754).

- need to revise names- are they both referring to the palea apiculus??

Moved 'glume opening' to glume morphology trait (TO:0000869)

Need to split lemma and palea color (TO:0000264) and lemma and palea pubescence (TO:0000417).