changed “amount” to “composition” in TO:0000196, TO:0000289
revised definitions of:

TO:0000297 - glucosamine content
old: Amount of glucosamine present in the plant or plant part

new: An oligosaccharide content (TO:0006003) which is the composition of glucosamine (CHEBI:5417) in a plant structure (PO:0009011).

TO_0020090>    'zinc content
old: The total zinc ion content measured in a plant or a plant part.

new: A mineral and ion content related trait (TO:0000465) which is the composition of zinc (CHEBI:27364) in a plant structure (PO:0009011).

TO:0020089 iron content-
old: The total iron ion content measured in a plant or a plant part.
new: A mineral and ion content related trait (TO:0000465) which is the composition of iron (CHEBI:24873) in a plant structure (PO:0009011).

TO_0020092>    'copper content'
old: The total copper ion content measured in a plant or a plant part.

new: A mineral and ion content related trait (TO:0000465) which is the composition of copper (CHEBI:23377) in a plant structure (PO:0009011).

TO_0000543>    'leaf nitrogen content'

old: A leaf composition trait (TO:0000493) which is the amount of nitrogen present in the leaf tissue in the form of minerals and metabolites.

new: A leaf composition trait (TO:0000493) which is the composition of nitrogen (CHEBI:51143) in the leaf (PO:0025034).

correct CHEBI ref added to each of the above terms.