oligosaccharide content (TO:0006003): changed equivilance axiom from 'amount' to 'composition'
manganese content (TO:0020091): updated def'n and chebi xref
old: The total manganese ion content measured in a plant or a plant part.
new: A mineral and ion content related trait (TO:0000465) which is the composition of manganese (CHEBI:25154) in a plant structure (PO:0009011).

phosphate content (TO:0020102): updated def'n
old: The total phosphate content measured in a plant or a plant part.
new: A mineral and ion content related trait (TO:0000465) which is the composition of phosphate (CHEBI:26020) in a plant structure (PO:0009011).

nitrate content (TO:0020094) updated def'n and chebi xref
old: The total nitrate content measured in a plant or a plant part.
new: A mineral and ion content related trait (TO:0000465) which is the composition of nitrate (CHEBI:17632) in a plant structure (PO:0009011).

abscisic acid content (TO:0002667)
changed chebi xref from 2-cis-absisic acid to the more generic 'abscisic acid' (CHEBI:62431)
updated the equivalence axioms to reflect this.

nitrogen content (TO:0020093)- updated def'n, and chebi xref
old: The total nitrogen content measured in a plant or a plant part.
new:  A mineral and ion content related trait (TO:0000465) which is the composition of nitrogen (CHEBI:51143) in a plant structure (PO:0009011).

neoxanthin content (TO:0020118)
old: The neoxanthin content in a plant or plant part.
new: A xanthophyll content (TO:0000301) which is the composition of neoxanthin (CHEBI:25501) in a plant structure (PO:0009011).

old: The zeaxanthin content in a plant or plant part.
new: A xanthophyll content (TO:0000301) which is the composition of zeaxanthin (CHEBI:27547) in a plant structure (PO:0009011).

violaxanthin content (TO:0020119)
old: The violaxanthin content in a plant or plant part.
new: A xanthophyll content (TO:0000301) which is the composition of violaxanthin (CHEBI:27295) in a plant structure (PO:0009011).

beta-cryptoxanthin (TO:0020079)
old: The beta-cryptoxanthin content in a plant or plant part.
new: A xanthophyll content (TO:0000301) which is the composition of beta-cryptoxanthin (CHEBI:10362) in a plant structure (PO:0009011).

Indole-3-acetic acid content (TO:0020106)
old: Measures the IAA content in a plant or plant part.
new: An auxin content (TO:0002672) which is the composition of indole-3-acetic acid (CHEBI:16411) in a plant structure (PO:0009011).

auxin content (TO:0002672)
old: Measures the auxin content in a plant or plant part.
new: A growth hormone content (TO:0000476) which is the composition of auxin (CHEBI:22676) in a plant structure (PO:0009011).

growth hormone content (TO:0000476)
old: Measures the growth hormone content in a plant or plant part.
new: A metabolite content trait (TO:0000281) which is the composition of plant hormone (CHEBI:37848) in a plant structure (PO:0009011).

carbohydrate derivative content (TO:0006008)
old: Total amount of carbohydrate derivatives present in the plant or plant part.
new: A carbohydrate content (TO:0000291) which is the composition of carbohydrate derivatives (CHEBI:63299) in a plant structure (PO:0009011).