Minor edits- Biochem branch: amylose content  (TO:0000196): slightly revised def'n: was: A polysaccharide content trait (TO:0006007) which is the composition of amylose (CHEBI:28102) present in a plant structure (PO:0009011). Revised: A polysaccharide content trait (TO:0006007) which is the amount of amylose (CHEBI:28102) in a plant structure (PO:0009011).
carotene content (TO:0000289):
Was:
A carotenoid content trait (TO:0000496) which is the composition of carotene  (CHEBI:23042) present in a plant structure (PO:0009011).

Revised:  A carotenoid content trait (TO:0000496) which is the amount of carotene (CHEBI:23042) in a plant structure (PO:0009011).

flavonoid content TO:0000290:
Revised:  A phytochemical compound content trait (TO:0000674) which is the amount of flavonoid (CHEBI:47916) in a plant structure (PO:0009011).

glucosamine content (TO:0000297):
An oligosaccharide content (TO:0006003) which is the amount of glucosamine (CHEBI:5417) in a plant structure (PO:0009011).
replaced DbxRefs