revised definitions, added correct CHEBI cross reference, and updated equivalence axioms to 'plant structure' instead of 'whole plant'
TO_0006053>    'zinc concentration'
old: The total Zinc ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A zinc content (TO:0020090) which is the concentration of zinc (CHEBI:27364) in some plant structure (PO:0009011)

TO_0006049>    'iron concentration'
old: The total iron concentration measured in a given volume of a plant or a plant part or plant extract.
new:  An iron content (TO:0020089) which is the concentration of iron (CHEBI:24873) in some plant structure (PO:0009011)

TO_0006043>    'boron concentration'
old: The total Boron ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of boron (CHEBI:22916) in some plant structure (PO:0009011).

TO_0006050>    'cobalt concentration'
old: The total Cobalt ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of cobalt (CHEBI:33888) in some plant structure (PO:0009011).

TO_0006052>    'copper concentration'
old: A copper content (TO:0020092) which is the amount of copper in a measured volumn.
new:  A copper content (TO:0020092) which is the concentration of copper (CHEBI:23377) in some plant structure (PO:0009011)

TO_0006054>    'arsenic concentration'
old: The total Arsenic ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of arsenic (CHEBI:22632) in some plant structure (PO:0009011).

TO_0006059>    'cadmium concentration'
old: The total Cadmium concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of cadmium (CHEBI:22978) in some plant structure (PO:0009011).

TO_0006047>    'calcium concentration'
old: The total Calcium ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of calcium (CHEBI:22985) in some plant structure (PO:0009011).

TO_0006046>    'sulfur concentration'
old: The total Sulfur concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of sulfur (CHEBI:26835) in some plant structure (PO:0009011).

TO_0006055>    'selenium concentration'
old: The total Selenium ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of salenium (CHEBI:26628) in some plant structure (PO:0009011).

TO_0006044>    'magnesium concentration'
old: The total Magnesium ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of magnesium (CHEBI:25108) in some plant structure (PO:0009011).

TO_0006048>    'manganese concentration'
old: The total Manganese ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A manganese content (TO:0020091) which is the concentration of manganese (CHEBI:25154) in some plant structure (PO:0009011)

TO_0006056>    'molybdenum concentration'
old: The total Molybdenum  concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of molybdenum (CHEBI:25370) in some plant structure (PO:0009011).

TO_0006045>    'phosphorus concentration'
old: The total Phosphorus concentration measured in a given volume of a plant or a plant part or plant extract.
new: A phosphorus content (TO:0001024) which is the concentration of phosphorus (CHEBI:26082) in some plant structure (PO:0009011)

TO_0006042>    'lithium concentration'
old: The total Lithium ion concentration measured in a given volume of a plant or a plant part or plant extract.
new: A mineral and ion content related trait (TO:0000465) which is the concentration of lithium (CHEBI:33298) in some plant structure (PO:0009011).