Revised names and def'ns:  all  have TO_GIT:105; TO:cooperl
* floret morphology trait: (TO:0000274):
was: A flower morphology trait (TO:0000499) which is a quality of a grass floret and its parts.
Revised:
spikelet floret morphology trait (TO:0000274): A flower morphology trait (TO:0000499) which is a quality of a spikelet floret (PO:0009082).

Child terms:
was:
- floret color (TO:0000536): A floret morphology trait (TO:0000274) which is the color of the floret.
Revised:
spikelet floret color (TO:0000536): A spikelet floret morphology trait (TO:0000274) which is the color of a spikelet floret (PO:0009082).

Was:
- floret diameter (TO:0000855): A floret morphology trait (TO:0000274) which is the diameter of a floret.

Revised:
spikelet floret diameter (TO:0000855): A spikelet floret morphology trait (TO:0000274) which is the diameter of a spikelet floret (PO:0009082).

Was:
- floret shape (TO:0000853): A floret morphology trait (TO:0000274) which is the shape of a floret.

Revised: spikelet floret shape (TO:0000853): A spikelet floret morphology trait (TO:0000274) which is the shape of a spikelet floret (PO:0009082).

Was:
	- floret size (TO:0000854): A floret morphology trait (TO:0000274) which is the size of a floret.

Revised: spikelet floret size (TO:0000854): A spikelet floret morphology trait (TO:0000274) which is the size of a spikelet floret (PO:0009082).

- floret number (TO:0000670)  -  moved to be is_a flower number trait, needs to be revised