New grain protein term and revisions to protein contents
Revised:
id: TO:0000598
name: protein content
def: "The total protein content measured in a plant or a plant part." [Gramene:pankaj_jaiswal37]
New: A metabolite content trait (TO:0000281) which is the amount of protein (CHEBI:36080) in a plant structure (PO:0009011).

Revised:
seed protein composition trait (TO:0000490): A seed composition based quality trait (TO:0000488) which is the amount of protein (CHEBI_36080) in a seed (PO:0009010). [source: unknown]

Revised: seed protein content  (TO:0000490): A protein content trait (TO:0000598) which is the amount of protein (CHEBI:36080) in a seed (PO:0009010).

New:
- grain protein content TO:0001136: A protein content trait (TO:0000598) which is the amount of protein (CHEBI:36080) in a caryopsis fruit (PO:0030104).

related synonyms:
- kernel protein content (related)
Revised:
* TO:0000410 polished rice protein content
Revised def'n: A grain protein content (TO:0001136) which is the amount of protein (CHEBI:36080) in a polished rice grain (caryopsis fruit; PO:0030104).

* TO:0000414 white rice protein content
NEW def'n: A grain protein content (TO:0001136) which is the amount of protein (CHEBI:36080) in a white rice grain (caryopsis fruit; PO:0030104).

* TO:0000138 brown rice protein
Revised def'n: A grain protein content (TO:0001136) which is the amount of protein (CHEBI:36080) in a brown rice grain (caryopsis fruit; PO:0030104).